[refactor] WBJ-572 Create team page components refactored & prettier applied
